updated hero carousel
