+++ src/aflglobal/src/assets/sass/abstracts/_mixins.scss
@@ -5,12 +5,14 @@
 $container-large: 1366px;
 $container: 1126px;
 $break-mobile: 380px;
+$break-mobile-medium: 399px;
 $break-mobile-large: 478px;
 $break-mobile-horizontal: 767px;
 $break-tablet: 900px;
 $break-tablet-horizontal: 991px;
 $break-tablet-large: 1199px;
 $max-width: 960px;
+$min-width: 251px;
 
 @mixin wrapper() {
     max-width: $max-width;
