+++ src/aflglobal/src/components/ContentBlock.tsx
@@ -16,7 +16,6 @@ type ContentBlockProps = ComponentProps & {
 const ContentBlock = ({ fields }: ContentBlockProps): JSX.Element => (
   <div className="contentBlock">
     <Text tag="h2" className="contentTitle" field={fields.heading} />
-
     <RichText className="contentDescription" field={fields.content} />
   </div>
 );
